refactor(dashboard): extract duplicated utilities into shared modules
- Create lib/constants.ts with unified ROLE_STYLES (better dark mode support)
- Create lib/format.ts with formatDate, formatDateShort, timeAgo
- Make ThemeToggle flexible with size/className props, replace inline SVGs with lucide icons
- Remove duplicated ThemeToggleIcon from app-sidebar, import shared ThemeToggle
- Remove duplicated ROLE_COLORS from project/page.tsx, ROLE_STYLES from conversations/page.tsx
- Remove duplicated date formatters from conversations/page.tsx and recent-activity.tsx
