+++ packages/dashboard/src/components/analytics/recent-activity.tsx
@@ -1,4 +1,5 @@
 import { MessageSquareIcon } from "lucide-react";
+import { timeAgo } from "@/lib/format";
 
 interface RecentConversation {
   id: string;
@@ -10,17 +11,6 @@ interface RecentConversation {
 
 interface RecentActivityProps {
   conversations: RecentConversation[];
-}
-
-function timeAgo(ts: number): string {
-  const seconds = Math.floor((Date.now() - ts) / 1000);
-  if (seconds < 60) return "just now";
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes}m ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h ago`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ago`;
 }
 
 export function RecentActivity({ conversations }: RecentActivityProps) {
