+++ packages/dashboard/src/lib/format.ts
@@ -1,0 +1,35 @@
+/**
+ * Date formatting utilities shared across the dashboard.
+ */
+
+/** Full date+time: "Jan 5, 2025, 02:30 PM" */
+export function formatDate(ts: number): string {
+  return new Date(ts).toLocaleString(undefined, {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+    hour: "2-digit",
+    minute: "2-digit",
+  });
+}
+
+/** Short date only: "Jan 5, 2025" */
+export function formatDateShort(ts: number): string {
+  return new Date(ts).toLocaleDateString(undefined, {
+    month: "short",
+    day: "numeric",
+    year: "numeric",
+  });
+}
+
+/** Relative time: "just now", "5m ago", "3h ago", "2d ago" */
+export function timeAgo(ts: number): string {
+  const seconds = Math.floor((Date.now() - ts) / 1000);
+  if (seconds < 60) return "just now";
+  const minutes = Math.floor(seconds / 60);
+  if (minutes < 60) return `${minutes}m ago`;
+  const hours = Math.floor(minutes / 60);
+  if (hours < 24) return `${hours}h ago`;
+  const days = Math.floor(hours / 24);
+  return `${days}d ago`;
+}
